Updated to use github_api_toolkit over local api interface. Now uses env variables for github app client id. Also gets .pem straight from AWS during runtime
