+++ pyproject.toml
@@ -13,6 +13,7 @@ flask = "^3.0.2"
 python-dateutil = "^2.9.0.post0"
 jwt = "^1.3.1"
 boto3 = "^1.34.118"
+github-api-toolkit = {git = "https://github.com/ONS-Innovation/github-api-package.git"}
 
 
 [build-system]
