+++ repoarchivetool/test_data/test_repositories.json
@@ -5,7 +5,7 @@
         "contributors": [],
         "apiUrl": "https://api.github.com/repos/ONS-Innovation/KPArchiveTest",
         "lastCommit": "2024-04-04",
-        "dateAdded": "2024-06-10",
+        "dateAdded": "2024-06-25",
         "exemptUntil": "1900-01-01",
         "exemptReason": "",
         "exemptBy": {
@@ -19,7 +19,7 @@
         "contributors": [],
         "apiUrl": "https://api.github.com/repos/ONS-Innovation/KPArchiveTest2",
         "lastCommit": "2024-04-04",
-        "dateAdded": "2024-06-12",
+        "dateAdded": "2024-06-27",
         "exemptUntil": "1900-01-01",
         "exemptReason": "",
         "exemptBy": {
@@ -33,8 +33,8 @@
         "contributors": [],
         "apiUrl": "https://api.github.com/repos/ONS-Innovation/KPInternalArchiveTest",
         "lastCommit": "2024-04-04",
-        "dateAdded": "2024-05-27",
-        "exemptUntil": "2024-10-09",
+        "dateAdded": "2024-06-11",
+        "exemptUntil": "2024-10-24",
         "exemptReason": "This repository is used to support other services within our team.",
         "exemptBy": {
             "name": "Kieran Pritchard",
@@ -47,7 +47,7 @@
         "contributors": [],
         "apiUrl": "https://api.github.com/repos/ONS-Innovation/KPPrivateArchiveTest",
         "lastCommit": "2024-04-04",
-        "dateAdded": "2024-06-10",
+        "dateAdded": "2024-06-25",
         "exemptUntil": "1900-01-01",
         "exemptReason": "",
         "exemptBy": {
